+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
-All notable changes to the "vue-support" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+All notable changes to the "vuejs-autoimport" extension will be documented in this file.
 
-## [Unreleased]
+## 0.0.2 | 2018-10-30
+
+- Changed to key command
+- Fixed to key for get config value
+- FIxed to import path on same layer
+
+## 0.0.1 | 2018-10-28
+
 - Initial release